refactor ghost circuit SH
