+++ src/main/java/gregtech/api/mui/widget/GhostCircuitSlotWidget.java
@@ -3,12 +3,13 @@ package gregtech.api.mui.widget;
 import gregtech.api.capability.impl.GhostCircuitItemStackHandler;
 import gregtech.api.mui.GTGuiTextures;
 import gregtech.api.mui.GTGuis;
-import gregtech.api.mui.synchandler.GhostCircuitSyncHandler;
 import gregtech.api.recipes.ingredients.IntCircuitIngredient;
 import gregtech.client.utils.TooltipHelper;
 
 import net.minecraft.item.ItemStack;
+import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraftforge.items.IItemHandler;
 
 import com.cleanroommc.modularui.api.drawable.IKey;
 import com.cleanroommc.modularui.api.widget.IWidget;
@@ -23,11 +24,14 @@ import com.cleanroommc.modularui.widgets.layout.Grid;
 import com.cleanroommc.modularui.widgets.slot.ModularSlot;
 import org.jetbrains.annotations.NotNull;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
 public class GhostCircuitSlotWidget extends ItemSlot {
+
+    private static final int SYNC_CIRCUIT_INDEX = 10;
 
     public GhostCircuitSlotWidget() {
         tooltipBuilder(this::getCircuitSlotTooltip);
@@ -111,7 +115,7 @@ public class GhostCircuitSlotWidget extends ItemSlot {
                         .background(GTGuiTextures.SLOT, new ItemDrawable(
                                 IntCircuitIngredient.getIntegratedCircuit(index)).asIcon())
                         .onMousePressed(mouseButton -> {
-                            getSyncHandler().syncToServer(10, buf -> buf.writeShort(index));
+                            getSyncHandler().syncToServer(SYNC_CIRCUIT_INDEX, buf -> buf.writeShort(index));
                             circuitPreview.setItem(IntCircuitIngredient.getIntegratedCircuit(index));
                             return true;
                         }));
@@ -129,5 +133,78 @@ public class GhostCircuitSlotWidget extends ItemSlot {
                         .matrix(options)
                         .minColWidth(18).minRowHeight(18)
                         .minElementMargin(0, 0)));
+    }
+
+    private static class GhostCircuitSyncHandler extends ItemSlotSH {
+
+        public GhostCircuitSyncHandler(ModularSlot slot) {
+            super(slot);
+        }
+
+        @Override
+        protected void phantomClick(MouseData mouseData) {
+            if (mouseData.mouseButton == 0) {
+                // increment on left-click
+                setCircuitValue(getNextCircuitValue(1));
+            } else if (mouseData.mouseButton == 1 && mouseData.shift) {
+                // clear on shift-right-click
+                setCircuitValue(GhostCircuitItemStackHandler.NO_CONFIG);
+            } else if (mouseData.mouseButton == 1) {
+                // decrement on right-click
+                setCircuitValue(getNextCircuitValue(-1));
+            }
+        }
+
+        @Override
+        protected void phantomScroll(MouseData mouseData) {
+            setCircuitValue(getNextCircuitValue(mouseData.mouseButton));
+        }
+
+        private void setCircuitValue(int value) {
+            GhostCircuitItemStackHandler handler = getGhostCircuitHandler();
+            handler.setCircuitValue(value);
+            syncToClient(1, buf -> {
+                buf.writeBoolean(false);
+                buf.writeItemStack(handler.getStackInSlot(0));
+                buf.writeBoolean(false);
+            });
+        }
+
+        @Override
+        public void readOnServer(int id, PacketBuffer buf) throws IOException {
+            if (id == SYNC_CIRCUIT_INDEX) {
+                setCircuitValue(buf.readShort());
+            } else {
+                super.readOnServer(id, buf);
+            }
+        }
+
+        private int getNextCircuitValue(int delta) {
+            GhostCircuitItemStackHandler handler = getGhostCircuitHandler();
+
+            // if no circuit, skip 0 and return 32 if decrementing,
+            // or, skip 0 and return 1 when incrementing
+            if (!handler.hasCircuitValue()) {
+                return delta == 1 ? 1 : IntCircuitIngredient.CIRCUIT_MAX;
+                // if at max, loop around to no circuit
+            } else if (handler.getCircuitValue() + delta > IntCircuitIngredient.CIRCUIT_MAX) {
+                return GhostCircuitItemStackHandler.NO_CONFIG;
+                // if at 1, skip 0 and return to no circuit
+            } else if (handler.getCircuitValue() + delta < 1) {
+                return GhostCircuitItemStackHandler.NO_CONFIG;
+            }
+
+            // normal case: change by "delta" which is either 1 or -1
+            return handler.getCircuitValue() + delta;
+        }
+
+        public GhostCircuitItemStackHandler getGhostCircuitHandler() {
+            IItemHandler handler = getSlot().getItemHandler();
+            if (!(handler instanceof GhostCircuitItemStackHandler ghostHandler)) {
+                throw new IllegalStateException(
+                        "GhostCircuitSyncHandler has IItemHandler that is not GhostCircuitItemStackHandler");
+            }
+            return ghostHandler;
+        }
     }
 }
